in App: created a handle search function for the submit function of the search form and added it to the context.
in the search form, created an on submit function for the form to call the context handle search and prevent default behavior.

in DailyForecast: added a conditional render.

in use mock fetch: removed the use effect hook, and passed in the state of the search box field.  also exported the get location function.